+++ NEWS.md
@@ -12,6 +12,7 @@
   even just models where compartment numbers are not compatible.
 * NMscanInput has the new argument recover.cols. Default is TRUE - use
   FALSE to not include columns that NONMEM did not read.
+* mergeCheck now retains column order from x.
 
 ## Bugfixes
 * NMwriteData now respects NMdataConf()$args.fwrite